database system in development

+++ src/main.js
@@ -7,6 +7,7 @@ const { Client, Collection, Intents, Message } = require('discord.js');
 const theWords = ['ploy','deep state','population control','coronavirus is fake','chinavirus','covid agenda', 'fake vaccine'];
 const tagDescriptions = ['spam','threat','misinformation']
 const dotenv = require('dotenv');
+const theInfo = ['https://www.cdc.gov/coronavirus/2019-ncov/index.html']
 dotenv.config();
 const client = new Client({ 
     intents: [
@@ -47,14 +48,14 @@ const sequelize = new Sequelize('database', 'user', 'password', {
 	// SQLite only
 	storage: 'database.sqlite',
 });
-const Tags = sequelize.define('tags', {
+const Tags = sequelize.define('tags', {//each instance (row) in the database is called a Tag for some reason
     name: {
         type: Sequelize.STRING,
         unique: true,
     },
     description: Sequelize.TEXT,
     username: Sequelize.STRING,
-    strikes: {
+    usage_count: {
         type: Sequelize.INTEGER,
         defaultValue: 1,
         allowNull: false,
@@ -63,9 +64,11 @@ const Tags = sequelize.define('tags', {
 
 client.once('ready', () => {
 	Tags.sync();
+    console.log('bot is up and running')
 });
 client.on('messageCreate', async msg =>{
     if (theWords.some(word => msg.content.includes(word))) {
+        //later on, this message will be modified to be dynanic and adaptive to type of malicious content detected
 		msg.reply(`${msg.author.username}, this is detected to be a malicious/disruptive comment. Please do not spread misinformation or spam content on the server. Here is information to educate:${theInfo[0]}`);
         const tagName = msg.author.id;
         const tagDescription = 'misinformation';
@@ -73,43 +76,47 @@ client.on('messageCreate', async msg =>{
 
         try {
             // equivalent to: INSERT INTO tags (name, description, username) values (?, ?, ?);
-            const tag = await Tags.create({
+            const tag = await Tags.create({//creates a database and 
                 name: tagName,
                 description: tagDescription,
                 username: Username,
-                strike: 1,
+                usage_count: 1,
             });
 
-            return interaction.reply(`Tag ${tag.name} added.`);
+            return msg.reply(`Tag ${tag.name} added.`);
         }
         catch (error) {
             if (error.name === 'SequelizeUniqueConstraintError') {//if user already added on there
-                const tagName = interaction.options.getString('name');//get the identifier for the tag and find it
                 const tag = await Tags.findOne({where: {name: tagName}});
                 if (tag) {
-                    tag.increment('strike')//records the amount of times user has been caught
+                    tag.increment('usage_count')//records the amount of times user has been caught
                     const addedDesc = await Tags.update({description: tagDescription + "\nmisinformation"}, {where: {name: tagName}});
                     if (!addedDesc) {//if the updating description process did not work...
                         console.log('something went wrong with updating')
                         return interaction.reply('not working at the moment')
                     }
-                    if (tag.get('strike') >= 3) {//suggests user be kicked if the strike # is or is greater than 3
-                        console.log(`${tag.get(username)} needs to be kicked. Please consider this decision.`)
+                    if (tag.get('usage_count') == 3) {//suggests user be kicked if the strike # is 3
+                        console.log(`${tag.get(Username)} needs to be kicked. Please consider this decision.`)
                     }
                 }
             }
             //or display the miscellaneous error
-            return interaction.reply(`Something went wrong with adding a tag. Error: ${error}`);
+            return msg.reply(`Something went wrong with adding a tag. Error: ${error}`);
         }
 	}
 })
 client.on('interactionCreate', async interaction => {
 	if (!interaction.isCommand()) return;
+
 	const { commandName } = interaction;
     if (commandName === 'discretion') {
+        const identityBuffer = await Tags.findOne({where: {name: interaction.user.username}});
+        if (identityBuffer) {//checks to see if the user isn't trying to get out of trouble
+            return interaction.reply("Sorry, but you cannot bail yourself out.")
+        }
 		const tagName = interaction.options.getString('username');
 
-        // equivalent to: DELETE from tags WHERE name = ?;
+        // takes the tag out of the system
         const rowCount = await Tags.destroy({ where: { name: tagName } });
 
         if (!rowCount) return interaction.reply('That tag did not exist.');
@@ -121,9 +128,6 @@ client.on('interactionCreate', async interaction => {
         console.log(`List of tags: ${tagString}`);
         return interaction.reply('Your request has been considered');
 
-	} else if (commandName === 'showtags') {
-		// equivalent to: SELECT name FROM tags;
-        
 	} 
 });
